RefGetParm return type changed from int to intptr_t for 64-bit
pointer compatibility. REF_API_VERSION bumped to 16.

Changes:
- engine/ref_api.h: updated RefGetParm signature and version
- ref/vk/vk_rmain.c: VK_RefGetParm return type to intptr_t
- ref/gl/gl_context.c: GL_RefGetParm return type to intptr_t
- ref/soft/r_context.c: GL_RefGetParm return type to intptr_t

+++ ref/vk/vk_rmain.c
@@ -227,7 +227,7 @@ static const char *getParmName(int parm)
 	}
 }
 
-static int VK_RefGetParm( int parm, int arg )
+static intptr_t VK_RefGetParm( int parm, int arg )
 {
 	// TODO all PARM_TEX handle in r_texture internally
 	switch(parm){
